+++ app/templates/es2015/src/scripts/app/services/templateFactory.js
@@ -1,0 +1,7 @@
+import mServices from './_mServices';
+
+mServices.factory('templateFactory', function () {
+    return {
+
+    };
+});