+++ app/templates/es5/src/scripts/app/directives/templateDirective.js
@@ -1,0 +1,13 @@
+'use strict';
+
+var mDirectives = require('./_mDirectives');
+
+mDirectives.directive('mdTemplateDirective', function () {
+    return {
+        restrict: 'AEC',
+        template: '',
+        link: function (scope, element) {
+            element;
+        }
+    };
+});