+++ app/templates/es5/src/scripts/app/services/templateFactory.js
@@ -1,0 +1,9 @@
+'use strict';
+
+var mServices = require('./_mServices');
+
+mServices.factory('templateFactory', function () {
+    return {
+
+    };
+});